+++ packages/api/src/resolvers/queries/events/events.ts
@@ -1,5 +1,5 @@
 import { log } from 'meteor/unchained:logger';
-import { Root, Context } from '@unchainedshop/types/api';
+import { Root, Context, SortOption } from '@unchainedshop/types/api';
 import { EventQuery } from '@unchainedshop/types/events';
 
 export default async function events(
@@ -14,7 +14,7 @@ export default async function events(
   }: EventQuery & {
     limit?: number;
     offset?: number;
-    sort: { created?: 'DESC' | 'ASC'; type?: 'DESC' | 'ASC' };
+    sort: Array<SortOption>;
   },
   { modules, userId }: Context,
 ) {